Create Theory-Assignment
What is Emmet?
Emmet is a free add-on for our text-editors. It allows us to type shortcuts that are then expanded into full piece of code. By using Emmet, developers type less, they save keystrokes and time.

What is CDN?
A Content Delivery Network(CDN) is a network of interconnected server that speed up web page loading for heavy-data applications.. When user visit a website, data from that website has to travel across the internet to reach the user's computer. If the user is locked far from the server, it will take time to load a large file, such as video or website images. Instead the website content is stored on CDN server geographically closer to the users and reaches their computers much faster.
Benefits:
a. Reduce page load time
b.Reduce bandwidth cost
c.Increase content availability
d. Improved website security

3. Why is react called as React?
It's called as React because it reacts. It was developed by FB (a site that constantly changes its data) to improve the user interface development and more effectively change (React To) what the user sees when they're doing things like mouse clicking, submitting and typing.

4. What is crossorigin in script tag?
The cross-origin attribute sets the mode of the request to an HTTP CORS Request. A cross-origin request is a request for a resource (eg. stylesheet, iframes, images, fonts or scripts) from another domain.

5. What is integrity in script tag?